docs(component): add doc for type lamda
affects: @action-land/component
