+++ modules/component/index.ts
@@ -1,3 +1,4 @@
 export {AutoForward} from './src/autoForward'
 export {Component, COM} from './src/component'
-export {ComponentNext, ComponentProps} from './src/componentNext'
+export {ComponentNext} from './src/componentNext'
+export {ComponentProps} from './types/componentProps'
