+++ modules/component/types/componentProps.ts
@@ -1,0 +1,11 @@
+import {Action} from '@action-land/core'
+
+export type ComponentProps = {
+  readonly iState?: unknown
+  readonly oState?: unknown
+  readonly iActions?: Action<unknown>
+  readonly oActions?: Action<unknown>
+  readonly oView?: unknown
+  readonly iChildren?: unknown
+  readonly iProps?: unknown
+}
