+++ modules/component/types/pickChildrenType.ts
@@ -1,0 +1,6 @@
+import {PPP} from './pickValueFromComponent'
+
+/**
+ * Picks component children type from component or component props
+ */
+export type iChildren<P> = PPP<P, 'iChildren'>
