+++ modules/component/types/pickIActionsType.ts
@@ -1,0 +1,6 @@
+import {PPP} from './pickValueFromComponent'
+
+/**
+ * Picks input actions type from component or component props
+ */
+export type iActions<P> = PPP<P, 'iActions'>
