+++ modules/component/types/pickIStateType.ts
@@ -1,0 +1,6 @@
+import {PPP} from './pickValueFromComponent'
+
+/**
+ * Picks input state type from component or component props
+ */
+export type iState<P> = PPP<P, 'iState'>
