+++ modules/component/types/pickOutputViewType.ts
@@ -1,0 +1,7 @@
+import {PPP} from './pickValueFromComponent'
+
+/**
+ * Picks input actions type from component or component props
+ */
+
+export type oView<P> = PPP<P, 'oView'>
