+++ modules/component/types/pickPropType.ts
@@ -1,0 +1,6 @@
+import {PPP} from './pickValueFromComponent'
+
+/**
+ * Picks component view props type from component or component props
+ */
+export type iProps<P> = PPP<P, 'iProps'>
