+++ modules/component/types/pickValue.ts
@@ -1,0 +1,4 @@
+/**
+ * Extracts values of the provided keys
+ */
+export type PP<O, K> = K extends keyof O ? O[K] : never
